added initalizers to all lists reports, finished presentation

+++ dev/DomainLayer/ProductFacade.java
@@ -20,7 +20,7 @@ public class ProductFacade
      * @param sellPrice the current sell price
      * @throws Exception if product already exists
      */
-    public void AddProduct(String productName, String category, List<String> subCategories, String manufacturer, int sellPrice, String shelfLocation, String storageLocation) throws Exception
+    public void AddProduct(String productName, String category, List<String> subCategories, String manufacturer, double sellPrice, String shelfLocation, String storageLocation) throws Exception
     {
         if(getProduct(productName)!=null)
             throw new Exception("Product with name: "+productName+" already exists!");
